+++ api/src/main/kotlin/com/nexters/bottles/api/global/resolver/RefreshTokenArgumentResolver.kt
@@ -1,7 +1,7 @@
 package com.nexters.bottles.api.global.resolver
 
+import com.nexters.bottles.api.auth.component.JwtTokenProvider
 import com.nexters.bottles.api.global.exception.UnauthorizedException
-import com.nexters.bottles.app.auth.service.JwtTokenProvider
 import org.springframework.core.MethodParameter
 import org.springframework.stereotype.Component
 import org.springframework.web.bind.support.WebDataBinderFactory
