+++ app/build.gradle.kts
@@ -4,11 +4,7 @@ dependencies {
 
     implementation("com.fasterxml.jackson.core:jackson-databind:2.15.0")
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
-
-    implementation("io.jsonwebtoken:jjwt-api:0.11.2")
-    implementation("io.jsonwebtoken:jjwt-impl:0.11.2")
-    implementation("io.jsonwebtoken:jjwt-jackson:0.11.2")
-
+    
     implementation("com.google.firebase:firebase-admin:9.2.0")
 
     implementation("io.github.microutils:kotlin-logging:3.0.5")
